refactor(beer-catalog): apply pre-push review fixes
Reconciled local pre-push review (Claude pr-pre-reviewer; the Codex helper
failed again on the --output-last-message CLI arg — noted, follow-up on
scripts/codex-review.sh).

- layering (Must Have): presentation no longer imports data/ — the named
  catalogue errors are re-exported through the application layer and the
  three detail screens import them from there.
- toNumber: empty/blank strings now map to null (not 0) + edge test.
- drop the unused `total` from the pagination hook's select (and from the
  screen-test factories) — no screen consumed it.
- computeNextPageParam made generic over Page<T> (brewery/style fast-follow
  reuse at zero runtime cost).
- import order: React first in BeerCard / CatalogListFooter.

Not changed (disagreed): the VM display labels (SOURCE_LABELS, founded
label) stay in application/beer-catalog.view-model.ts — moving them to
presentation/beer-catalog.constants.ts would invert the dependency rule
(application importing presentation), the same rule the Must Have enforces;
scan keeps its placeholders in a lower layer for the same reason.
