+++ packages/mobile-app/src/features/beer-catalog/presentation/BreweryDetailScreen.tsx
@@ -24,7 +24,7 @@ import {
   type BreweryFicheVM,
 } from "@/features/beer-catalog/application/beer-catalog.view-model";
 import { useBrewery } from "@/features/beer-catalog/application/use-beer-catalog-details";
-import { CatalogNotFoundError } from "@/features/beer-catalog/data/beer-catalog.errors";
+import { CatalogNotFoundError } from "@/features/beer-catalog/application/beer-catalog.use-cases";
 import {
   BACK_ACCESSIBILITY_LABEL,
   BACK_LABEL,
