+++ packages/mobile-app/src/features/beer-catalog/presentation/__tests__/BeerCatalogSearchScreen.test.tsx
@@ -51,7 +51,6 @@ type PaginationResult = ReturnType<typeof useBeerCatalogPagination>;
 
 interface PaginationOverrides {
   beers?: CatalogBeer[];
-  total?: number;
   isLoading?: boolean;
   isFetching?: boolean;
   isFetched?: boolean;
@@ -69,7 +68,6 @@ function buildPaginationResult(
 ): PaginationResult {
   const {
     beers = [],
-    total = beers.length,
     isLoading = false,
     isFetching = false,
     isFetched = true,
@@ -80,7 +78,7 @@ function buildPaginationResult(
   } = overrides;
 
   return {
-    data: isLoading ? undefined : { beers, total },
+    data: isLoading ? undefined : { beers },
     isLoading,
     isFetching,
     isFetched,
