+++ packages/mobile-app/src/features/beer-catalog/presentation/components/BeerCard.tsx
@@ -1,8 +1,7 @@
+import React from "react";
 import { Pressable, StyleSheet, Text, View } from "react-native";
 
 import { Ionicons } from "@expo/vector-icons";
-import React from "react";
-
 import { colors, radius, spacing, typography } from "@/core/theme";
 import { Card } from "@/core/ui/Card";
